+++ buildSrc/src/main/kotlin/Metadata.kt
@@ -1,0 +1,9 @@
+object Metadata {
+    const val GROUP_ID = "io.github.woody230"
+    const val SUBGROUP_ID = "ktx"
+    const val NAMESPACE_ID = "com.bselzer"
+    const val VERSION = "5.4.0"
+    const val JVM_TARGET = "11"
+    const val COMPILE_SDK = 33
+    const val MIN_SDK = 21
+}
